requirements.txt
actualpy==0.19.0
annotated-types==0.7.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dotenv==0.9.9
idna==3.11
proto-plus==1.26.1
protobuf==6.33.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.44
sqlmodel==0.0.27
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
